Split file into individuals
