+++ src/Deserializer/WPP/DecodingContextType.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.ObjectModel;
-using System.Diagnostics.CodeAnalysis;
 using System.Text;
 
 using Nefarius.Utilities.ETW.Deserializer.WPP.TMF;
@@ -21,7 +20,7 @@ public abstract class DecodingContextType
     protected ReadOnlyMemory<byte> Buffer { get; init; }
 
     /// <summary>
-    ///     Collection of extracted <see cref="TraceMessageFormat"/>s of this <see cref="DecodingContextType"/>.
+    ///     Collection of extracted <see cref="TraceMessageFormat" />s of this <see cref="DecodingContextType" />.
     /// </summary>
     internal abstract ReadOnlyCollection<TraceMessageFormat> TraceMessageFormats { get; }
 
@@ -49,64 +48,4 @@ public abstract class DecodingContextType
             };
         }
     }
-}
-
-/// <summary>
-///     A <see cref="TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFSEARCHPATH" />  wrapper for use with
-///     <see cref="DecodingContext" />.
-/// </summary>
-[SuppressMessage("ReSharper", "UnusedType.Global")]
-public sealed class TmfFilesDecodingContextType()
-    : DecodingContextType(TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFSEARCHPATH)
-{
-    /// <summary>
-    ///     Gets decoding info from multiple paths containing <c>.TMF</c> files.
-    /// </summary>
-    /// <param name="path">
-    ///     Null-terminated Unicode string that contains the path to the .tmf file. You do not have to
-    ///     specify this path if the search path contains the file. Only specify this context information if you also specify
-    ///     the TDH_CONTEXT_WPP_TMFFILE context type. If the file is not found, TDH searches the following locations in the
-    ///     given order:
-    ///     <ul>
-    ///         <li>The path specified in the TRACE_FORMAT_SEARCH_PATH environment variable</li>
-    ///         <li>The current folder</li>
-    ///     </ul>
-    /// </param>
-    public TmfFilesDecodingContextType(string path) : this()
-    {
-        ArgumentNullException.ThrowIfNull(path);
-        Buffer = new ReadOnlyMemory<byte>(Encoding.Unicode.GetBytes(Path.GetFullPath(path))
-            .Concat("\0\0"u8.ToArray())
-            .ToArray()
-        );
-    }
-
-    /// <inheritdoc />
-    internal override ReadOnlyCollection<TraceMessageFormat> TraceMessageFormats => throw new NotImplementedException();
-}
-
-/// <summary>
-///     A <see cref="TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFFILE" />  wrapper for use with <see cref="DecodingContext" />.
-/// </summary>
-[SuppressMessage("ReSharper", "UnusedType.Global")]
-public sealed class TmfFileDecodingContextType() : DecodingContextType(TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFFILE)
-{
-    /// <summary>
-    ///     Gets decoding info from a single <c>.TMF</c> file.
-    /// </summary>
-    /// <param name="path">
-    ///     Null-terminated Unicode string that contains the name of the .tmf file used for parsing the WPP log.
-    ///     Typically, the .tmf file name is picked up from the event GUID so you do not have to specify the file name.
-    /// </param>
-    public TmfFileDecodingContextType(string path) : this()
-    {
-        ArgumentException.ThrowIfNullOrEmpty(path);
-        Buffer = new ReadOnlyMemory<byte>(Encoding.Unicode.GetBytes(Path.GetFullPath(path))
-            .Concat("\0\0"u8.ToArray())
-            .ToArray()
-        );
-    }
-
-    /// <inheritdoc />
-    internal override ReadOnlyCollection<TraceMessageFormat> TraceMessageFormats => throw new NotImplementedException();
 }
